Set x-speakeasy-unknown-values for non-boolean enums
Speakeasy generates "closed enums" by default, which doesn't work well
for forward compatibility, meaning that when Plex updates their API and
the response types include a new type of value, it can result in errors
when unmarshaling the response in one of the generated SDKs.

To resolve this, they support using "open enums" by specifying the
x-speakeasy-unknown-values on the schema for an enum.

For details see https://www.speakeasy.com/docs/customize/data-model/enums#open-vs-closed-enums

I've only does this for non-boolean enums, ie: enums where the values
aren't generally 0 or 1 for the purposes of true/false.

I also did this for non-model schemas (paths) too. This allows
users of the SDK to manually provide a value for an unsupported enum in
their request parameters. This is useful when the SDK hasn't been
updated to support a new value available for filtering for example.

Fixes https://github.com/LukeHagar/plex-api-spec/issues/95

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>